refactor(cmd): extract Record + recordFrom to record.go
output.go had grown to 829 LOC with the 410-line Record type + its
recordFrom projection sitting alongside 23 print/format functions.
Splits the wire-shape concerns into record.go and leaves output.go
holding only the print logic. Mirrors the per-tool split shipped in
#84 for internal/mcpserver. Pure mechanical refactor: no behaviour,
no public API, no test changes.

  output.go  829 → 418 LOC
  record.go  (new) 417 LOC
